新形式のNPCバトルコースを追加した
