+++ src/js/game/actions/game-actions.js
@@ -2,8 +2,8 @@
 
 import type {ArmDozerId, GameEnd, PilotId} from "gbraver-burst-core";
 import type {PostNetworkError} from '../network/post-network-error';
-import type {NPCBattleCourseDifficulty} from "../npc-battle/npc-battle-course";
 import type {GbraverBurstBrowserConfig} from "../config/browser-config";
+import type {NPCBattleCourseDifficulty} from "../npc-battle-course-master";
 
 /** 画面リロード依頼 */
 export type ReloadRequest = {
