+++ src/js/game/dom-dialogs/difficulty/difficulty-dialog.js
@@ -3,7 +3,6 @@ import {Howl} from 'howler';
 import type {Resources} from "../../../resource";
 import type {DOMDialog} from '../dialog';
 import {PathIds} from "../../../resource/path";
-import type {NPCBattleCourseDifficulty} from "../../npc-battle/npc-battle-course";
 import type {Stream, StreamSource, Unsubscriber} from "../../../stream/core";
 import {RxjsStreamSource} from "../../../stream/rxjs";
 import {domUuid} from "../../../uuid/dom-uuid";
@@ -12,6 +11,7 @@ import {pushDOMStream} from "../../../dom/push/push-dom";
 import type {PushDOM} from "../../../dom/push/push-dom";
 import {pop} from "../../../dom/animation/pop";
 import {SOUND_IDS} from "../../../resource/sound";
+import type {NPCBattleCourseDifficulty} from "../../npc-battle-course-master";
 
 /** ルート要素 class属性 */
 const ROOT_CLASS = 'difficulty';
