+++ src/js/game/npc-battle/npc-battle-course.js
@@ -3,9 +3,7 @@
 import type {ArmDozerId} from "gbraver-burst-core";
 import type {StageLevel} from "./npc-battle-stage";
 import type {NPCBattleStage} from "../npc-battle";
-
-/** NPCバトルコース難易度 */
-export type NPCBattleCourseDifficulty = 'Easy' | 'Normal' | 'Hard';
+import type {NPCBattleCourseDifficulty} from "../npc-battle-course-master";
 
 /** NPCバトルコース */
 export interface NPCBattleCourse {
